Refactor project structure and update dependencies
- Updated `libconfig` subproject commit.
- Modified `CMakeLists.txt` to include additional libraries.
- Commented out frame management calls in `editor.cpp`.
- Changed file inclusions to reflect new structure across multiple files.
- Introduced new files for project management: `project.cpp`, `project.h`, `project_settings.cpp`, and `project_settings.h`.
- Enhanced error handling in `ErrMsg` for cross-platform compatibility.
- Updated `CreateBuffer` function to accept `uint64_t` for size.
- Refactored `CommandBuffer` constructor and added new methods in `GraphicsEngine`.
- Introduced `FileSystem` class for improved file management.

+++ source/GraphicsEngine/vulkan/vk_buffers.h
@@ -47,7 +47,7 @@ namespace SceneryEditorX
      * allocating it in the appropriate memory type based on the memory parameter.
      * It abstracts away the details of buffer creation and memory allocation in Vulkan.
      */
-    Buffer CreateBuffer(uint32_t size, BufferUsageFlags usage, MemoryFlags memory = MemoryType::GPU, const std::string& name = "");
+    Buffer CreateBuffer(uint64_t size, BufferUsageFlags usage, MemoryFlags memory = MemoryType::GPU, const std::string& name = "");
     
     /**
      * @brief Copies data between two Vulkan buffers
